Stop FX before render
Rewinding cursor to the beginning was not helping lingering FX. The internet seems to think the preference in this commit is a more common way to force all VSTi (e.g. reverbs) to stop emitting sound. Disabling the preference and minimizing the tail work with manual cursor moving and playback testing. Hopefully it works with render too.

* Require SWS Extension, as it can get/set Reaper preferences
* Remove ineffective rewind cursor