Revert "Force media items online before render"
This commit doesn't do what it says. The preference used only forces media items _offline_ when Reaper is in the background. The preference never forces media items online. If the preference starts set and this project runs with Reaper in the background, there seems to be no API to force them online.

This reverts commit 884805613817ff9790dffbb078f53e481e4cdfa3.