Switch render processing to async functions
Continues the work of #4, to convert render to async. The previous work only converted the Click command to async. Now processing in this project is async, from top to bottom. There is still no speed benefit, as, at the bottom, is Reaper's Python API, which is sync, not async. This change prepares for the day.

* Encapsulate the existing Reaper render command call in a method on this project's `ExtendedProject` class