Switch definition of instrumental and DJ instrumental
The instrumental is now devoid of vocal samples, scratches, and sound effects. Meanwhile, the DJ instrumental has such samples, such that I could rap over it live, and it would be close to the published main song.

My music increasingly leans on vocal samples, scratches, and sound effects. As an instrumental listener, or as a rap remixer or freestyler, these effects are annoying interruptions. The published instrumental should be annoyance-free. Meanwhile, the DJ instrumental is the lesser distributed version, mainly for my performance or private enjoyment, so it can contain the annoying outtakes.

This is a breaking change for songs published prior, e.g. my Self Less (Instrumentals) album. However, rendering the instrumental version this new way is the direction I want to go for my next songs to publish.